requirements.txt
# Google Agent Development Kit
google-adk>=0.1.0
google-genai==1.0.0
google-cloud-aiplatform==1.73.0

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.11.0

# Data Science & ML
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.0
prophet==1.1.5

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# External APIs
requests==2.31.0

# Dashboard
streamlit==1.28.0
plotly==5.17.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
loguru==0.7.2

# For notebook development (Optional)
jupyter==1.0.0
ipykernel==6.25.0
holidays==0.35
